chore: update test proof for CI verification
