+++ artifacts/test-proof.enc
@@ -1,6 +1,6 @@
 {
-  "encryptedAesKey": "eKjrPvBTFX/PYh3sy636z1laYJ89hJYDuEZS3y1Qq3s4w3POpRy7YLZNOh/jBuxA/sEb735hHtXP9n0c9qugdABBa6rRibJiRG0fGnM15uJL8rmobmD5B2Rr7EDyFzLTSIKb1UI8mP5+hu1ssQ6KMtOfqj5mZtKTA5NHPP/zjJnJiwCBXAYCKcB39uutEXAKUOJyaduzN4rslUOPcUzEENYzkM8ZmaRfATfWFvuOCsTGT0yPWpfuH46O5A1gjkgiqCOXHX3Wb1pSNNY/r6qAmvGK86zYrt8Z/CZf6ntTl/zPQ/gT4rn3jA0jvz4/8nwcalMQx4oyoH0amNScXp8h6g==",
-  "iv": "KAiVVJ59ijlVGIWenyFQCA==",
-  "authTag": "I1JeyB0G2MU61D0x7qMB7A==",
-  "encryptedBody": "4E7WLu4buf96cB/fFR1T2FX8Ig5A68m54mIEYjdchQGw2UsnbK58vOX+eqnIEGToA+c23Sm8utuFTQAz5FD0Npayz0airmMMJWsq5x6+NaT++qnBG42BcdW51LCFLrDqo+PDwPVlEHY7guW9m/6gVrLT64Boq3HTu5XElIT82JD+3dVdq4CJ84MVoZHDauFn+UqUZhtbbAsEpwx/tY1z05I1Zsu/p2aXG0vsD1gJmtlnB6LEklp9RyM+I6aAj1k5atUd5F6zsMl9OYQNjPK+KvFry3/XvuFVkJLyVntv/ciJX0wOB+x70dRrR1eaPkjFOMBKMkpB2Gpc2bt9OFHZQPX5Lk2mQ5lqOOXD6cQsbp5nHrI76mJNEUhIb9Lgv0gw2/LkN/KgHQAX2umloREKh7rBV3Pc4WWjBVnvfc3R/+wEHFY1ezTnV2WR57qOvfNCKoZx3Ugl0TWjtYVvN7spAfYCcevw8z/5PAwFyPZHf/ufKk7P/lkoIOn7F580v7cpxMw2lwCtTUxbg/NbRuW4lEyxSTCvy2T9uLv8O/45ePmAdbZelkbXGx6P5A=="
+  "encryptedAesKey": "dH9KB23nHuCprhKR8y5Qp5PWCMtOIb5ZjO8i8gfrPXFksbXHbs3qXKLxSCfCD2G/pJp2+/UW6/5i/1M5IiWl5agL/fTMuyfAEhQ8zT2wK2p/TJoq/5Xg3SxKmdqGzfaWGe+IFICpW22wV/Tbg8jpvf04oFq3Bnq7tfq/3qe8IKCy4wlArgZfbuT8Hsq66VdH+kdUrDqosJAv3NSFMk2C/VRzQqDBcQ/uKWuD/Iq92y1xjiml7zGDdfHvPvwzJuLB+ETqME+KdoPD8tiD1PNqdU/E8MZjKKkD51mCS6gO7UpmgXEwhtRMzUaElH4JUi7K7A2Qy6uiV/aYs8Bt/6h7ew==",
+  "iv": "8pSvD/zWPDG9IIRVpW4mdg==",
+  "authTag": "Tdp6adHC1BUb4ya4ItV+oQ==",
+  "encryptedBody": "YwyKZHHcDUS+6/3vGMvL2eZL3/1raz9K0ZYq5LIEUkx5dDWHGlrpIfBjWMzWEYSDaaw87V7FPYGIkxWW55PSH6hfc1Dvx79yqcp42deHGMiO7VPl2suowpGjBLffE14O5deNP2Tt8Ch+DZPg5/ijyc4Wb2sVCbqu5MnylZc0Skf+K68RgWJlJbUHQd8LkBLshJqi2fsXpgTwRP3IqE9SI525pqOBO5o4Pw9ARLiuOOUYwRwrf1qkPSqDvtlMumqKq6q8Tupm25nAvvaBPns+tSger8+W6KoCcbA1eEhoyFMt4bD56lL85EZe3jY0BAfu+m+mwhfjdjLGHUHslPz4HP/fWSihlBFtDGpH13aIOo6Y4WwCCF0xX//TsUXx/BcqrRn4GuVj02/OG40BsEuPfjnBISTphEnmo/TCx0/MMBNu82rlRsS6XRqtLQdkktQPtjMhxlDaMJi7aYXahsKMLpojqQC2SvcTh1+qOuh9n18gOKQlwXRpokahlL3uDxI2t/aEP5rFMASug1WA8cpmR7ZLxXrkZ/VEQ/XoFefFIg/rumKPaVJFQ8gLBGdCmQ5dnIkUqP7FZvOiRxmNQcxrzEwBiZgkxg=="
 }
